pyproject.toml
[project]
name = "movietracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
